MFC r349898, r349916:
ipfilter commands, in this case ipf(8), passes its operations and rules
via an ioctl interface. Rules can be added or removed and stats and
counters can be zeroed out. As the ipfilter interprets these
instructions or operations they are stored in an integer called
addrem (add/remove). 0 is add, 1 is remove, and 2 is clear stats and
counters. Much of this is not documented. This commit documents these
operations by replacing simple integers with a self documenting
enum along with a few basic comments.


git-svn-id: https://svn.freebsd.org/base/stable/11@350110 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f