MFC r344226, r344234: stand: zfs memory corruption bug
r344226:
Fix memory corruption bug introduced in r325310

The bug occurred when a bounce buffer was used and the requested read
size was greater than the size of the bounce buffer.  This commit also
rewrites the read logic so that it is easier to systematically verify
all alignment and size cases.

r344234:
It turns out r344226 narrowed the overrun bug but did not eliminate it entirely

This commit fixes a remaining output buffer overrun in the
single-sector case when there is a non-zero tail.


git-svn-id: https://svn.freebsd.org/base/stable/12@350341 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f