MFC r350228:
ata_xpt: Use the correct union member when accessing valid.

In principle this should not matter as it's a union and they point to
the same memory location but based on the code above we should be
accessing .sata and not .ata.

Submitted by:	arichardson
Reviewed by:	scottl, imp
Obtained from:	CheriBSD
Sponsored by:	DARPA, AFRL
Differential Revision:	https://reviews.freebsd.org/D21002


git-svn-id: https://svn.freebsd.org/base/stable/12@350543 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f