bhyve: Ignore MSI/MSI-X interrupts sent to non-active vCPUs in physical destination mode.
This is mostly a nop, because the vmm initializes all vCPUs up to
vm_maxcpus, so even if the target CPU is not active, lapic/vlapic code
still has the valid data to use.  As John notes, dropping such
interrupts more closely matches the real harware, which ignores all
interrupts for not started APs.

Reviewed by:	jhb
admbugs:	837
MFC after:	1 week
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/head@350560 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f