objdump: update deprecation notice
MFC r350503: objdump: move deprecation notice to indended spot

r335217 added a deprecation notice to the source file for the objdump
man page, and r335219 added it to the rendered objdump.1, but in the
wrong spot.

MFC r350505: objdump: be explicit that GNU objdump that will be removed

We may install llvm-objdump as objdump (see review D18307) or just
provide no /usr/bin/objdump, but either way GNU objdump won't be
installed in the future.

Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/12@350565 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f