MFC r340987, r340989, r341254, r341269, r341333
r340987:
arm64: Add evdev support to GENERIC

r340989:
regulator_fixed: Do not disable fixed regulator at probe

If the regulator is unused it will be disabled by the regulator_shutdown sysinit.

Tested on pinebook where the backlight is controlled by a fixed-regulator.
The regulator doesn't have a regulator-boot-on param (I'm gonna upstream this) and so we disable it at probe.
We later enable it but this cause the screen to go black.
Linux doesn't disable regulator at boot (at least for fixed-regulator) so better match this to have the same UX.

Differential Revision:	https://reviews.freebsd.org/D17978

r341254:
extres: clk: Fix clk_set_assigned

ofw_bus_parse_xref_list_get_length doesn't returns the number of elements, fix this.
While here when setting the clock to the assigned freqeuncy, allow the clock
driver to round down or up the frequency as sometimes the exact frequency cannot
be obtain.

r341269:
release: arm64: Add opp dtbo to PINE* boards

r341333:
arm64: allwinner: Add 792Mhz frequency to sun50i-a64-opp

This is the frequency of the cpu on the Pinebook so add it to make
cpufreq find the current setting.
Note that this dtbo on the Pinebook doesn't work right now as u-boot
dtb doesn't have symbols and so it fails to apply. Linux 4.20 have
the dts and will be imported once taggued.

X-MFC with:	r341268


git-svn-id: https://svn.freebsd.org/base/stable/12@350593 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f