MFC r345948, r345951
r345948:
twsi: Add interrupt mode

Add the ability to use interrupts for i2c message.
We still use polling for early boot i2c transfer (for PMIC
for example) but as soon as interrupts are available use them.
On Allwinner SoC >A20 is seems that polling mode is broken for some
reason, this is now fixed by using interrupt mode.
For Allwinner also fix the frequency calculation, the one in the code
was for when the APB frequency is at 48Mhz while it is at 24Mhz on most
(all?) Allwinner SoCs. We now support both cases.

While here add more debug info when it's compiled in.

Tested On: A20, H3, A64

r345951:
twsi: Use config_intrhook_oneshot instead of config_intrhook_establish

Suggested by:	ian
X-MFC-With:	345948


git-svn-id: https://svn.freebsd.org/base/stable/12@350596 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f