MFC r344633-r344634, r344638
r344633:
usb_nop_xceiv: Add support for this pseudo device

This is a "fake" phy that handle regulator, clocks and reset gpio.
Only clock and regulator is supported for now.

Sponsored-by:      Rubicon Communications, LCC ("Netgate")

r344634:
xhci_mv: Move the driver to generic_xhci

Marvell XHCI is in fact generic-xhci, so move the driver and
add the compatible string.
While here, get and enable the phy if the dtb provide one.
The xhci bindings state that phys should be in a 'phys' property but
Marvell DTS uses 'usb-phy', only add support for 'usb-phy' for now.

Sponsored-by:      Rubicon Communications, LCC ("Netgate")

r344638:
Fix armv6/armv7 build after the move from xhci_mv to generic_xhci


git-svn-id: https://svn.freebsd.org/base/stable/12@350601 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f