cap_filergs: limit size of the file name
The limit of the name in fileargs is twice the size of the MAXPATH.
The nvlist will not add an element with the longer name.
We can detect at this point that the path is too big, and simple return
the same error as open(2) would.

PR:		239700
Reported by:	markj
Tested by:	markj
MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@350695 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f