MFC r344070 (by imp): Fix panic message.
The panic message lead people to believe some userland CAM request had
caused a problem when in reallity it was for a kernel request (eg the
USER bit was cleared). Reword message. Also, improve a couple of
comments to reflect that the periph shouldn't be completely torn down
before we get here (so the path and sim pointers should be valid, but
aren't and the code is designed to be robust enough in the face of
that to give a specific panic message).


git-svn-id: https://svn.freebsd.org/base/stable/12@350709 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f