MFC 346360: Push down INP_WLOCK slightly in tcp_ctloutput.
The inp lock is not needed for testing the V6 flag as that flag is set
once when the inp is created and never changes.  For non-TCP socket
options the lock is immediately dropped after checking that flag.
This just pushes the lock down to only be acquired for TCP socket
options.

This isn't a hot-path, more a cosmetic cleanup I noticed while reading
the code.


git-svn-id: https://svn.freebsd.org/base/stable/12@351057 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f