MFC r350881:
Calculate the number interface array elements using the new FR_NUM macro
instead of the hard-coded value of 4. This is a precursor to increasing
the number of interfaces speficied in "on {interface, ..., interface}".
Note that though this feature is coded in ipf_y.y, it is partially
supported in the ipfilter kld, meaning it does not work yet (and is yet
to be documented in ipf.5 too).


git-svn-id: https://svn.freebsd.org/base/stable/12@351479 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f