Fix LINT kernel builds on powerpc64 and sparc64.  This is a direct commit to 12-stable because the nandfs code no longer exists in 13-current.
The build was failing with

 nandfs_dat.c:301:
  warning: comparison is always false due to limited range of data type

I tried to fix it with an inline (size_t) cast of nargv->nv_nmembs in the
if() expression, but that didn't help (which seems a bit buggy), but using
an intermediate variable fixed it.  Elegance doesn't matter as much as
suppressing the warning; this code is long-dead even on this branch.


git-svn-id: https://svn.freebsd.org/base/stable/12@351877 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f