MFC r351747:
  Implement nvme suspend / resume for pci attachment

Note: this is merged ~9 hours early due to a desire to have it in before feature
freeze in 20 minutes. Several reports of it working in current (and one that it
worked in -stable after copying all of -current's nvme driver) gives me
confidence that bending the rules a little here is the right trade-off.

PR: 240340
Relnotes: Yes


git-svn-id: https://svn.freebsd.org/base/stable/12@351914 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f