MFC r351413,351459,351467: unbreak last(1) for 8-bit locales
Ouput format of last's broken for non UTF-8 locales
since it got libxo(3) support. It uses strftime(3) that produces
non UTF-8 strings passed to xo_emit(3) with wrong %s format -
it should be %hs in this case, so xo_emit(3) produces empty output.

This change is basically no-op when locale is of UTF-8 type,
f.e. en_GB.UTF-8 or ru_RU.UTF-8 or sr_RS.UTF-8@latin.
Also it is no-op for C/POSIX locale that's a subset of UTF-8.

It fixes output for other locales.


git-svn-id: https://svn.freebsd.org/base/stable/12@351924 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f