MFC r351797: (Optional) Allow more nesting of GEOM partitioning schemes
GEOM is supposed to be topology-agnostic, but the GPT and BSD partition code
has arbitrary restrictions on nesting that are annoying in cases such as
running VMs on raw partitions (since the VM's partitioning scheme is not
visible to the host).

This patch adds sysctls to disable the restrictions except in the case of
BSD label (and similar) partitions with offset 0 (where we need to avoid
recursively recognizing the label).


git-svn-id: https://svn.freebsd.org/base/stable/12@352084 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f