MFC r344555 (sobomax): makefs: Improve error handling
Bail out if one of the files scheduled to go to the FS image we are
making cannot be read (e.g. EPERM).  Previous behaviour of issuing a
warning but still proceeeding and returning success was definitely not
correct: masking out error condition as well as making a slighly
inconsistent FS where attempt to access the file in question ends up in
EBADF.  (More details in review D18584.)


git-svn-id: https://svn.freebsd.org/base/stable/12@352085 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f